navialife/requirements.txt
asgiref==3.4.1
backcall==0.2.0
blis==0.7.4
catalogue==1.0.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
cymem==2.0.5
decorator==5.1.0
Django==3.2.7
django-bulk-update-or-create==0.3.0
en-core-med7-lg==0.0.3
idna==3.2
importlib-metadata==4.8.1
ipython==7.27.0
jedi==0.18.0
matplotlib-inline==0.1.3
murmurhash==1.0.5
mysqlclient==2.0.3
numpy==1.21.2
pandas==1.3.3
parso==0.8.2
pickleshare==0.7.5
plac==1.1.3
preshed==3.0.5
prompt-toolkit==3.0.20
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
spacy==2.3.5
sqlparse==0.4.2
srsly==1.0.5
thinc==7.4.5
tqdm==4.62.2
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.2.5
zipp==3.5.0
